Sandardization of texture path.

+++ Nietsneflow3d/ECS/Systems/MapDisplaySystem.cpp
@@ -100,8 +100,8 @@ void MapDisplaySystem::fillVertexFromEntities()
                                                             Components_e::SPRITE_TEXTURE_COMPONENT);
         assert(posComp);
         assert(spriteComp);
-        assert(static_cast<size_t>(spriteComp->m_spriteData->m_textureNum) < m_vectVerticesData.size());
-        m_vectVerticesData[static_cast<size_t>(spriteComp->m_spriteData->m_textureNum)].
+        assert(spriteComp->m_spriteData->m_textureNum < m_vectVerticesData.size());
+        m_vectVerticesData[spriteComp->m_spriteData->m_textureNum].
                 loadVertexStandartTextureComponent(*posComp, *spriteComp);
     }
 }
